Support Pulseaudio "monitor" devices
In Pulseaudio, each output device has a corresponding input device.
These are actually useful, e.g. when listening to a received signal on
the headphones (perhaps from the websdr or some SDR software), it makes
sense to attach the HS modem to "Monitor of Headphones" to have it
decode it. Pulseaudio also supports virtual sinks that can be fed from
SDR software, and letting the HS modem listen on "Monitor of virtual
sink" will enable decoding.